requirements.txt
asgiref==3.9.2
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
coreapi==2.3.3
coreschema==0.0.4
Django==5.2.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
Faker==37.8.0
google-auth==2.40.3
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
openapi-codec==1.3.2
packaging==25.0
pillow==11.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
setuptools==80.9.0
simplejson==3.20.2
sqlparse==0.5.3
tzdata==2025.2
unique-namer==1.6.2
uritemplate==4.2.0
urllib3==2.5.0
